Extract common JS logic into a shared module
