+++ login/index.html
@@ -6,7 +6,7 @@
   <link rel="stylesheet" href="/css/common.css">
   <link rel="stylesheet" href="/css/auth.css">
   <link rel="stylesheet" href="/css/responsive.css">
-  <script src="/js/auth.js" defer></script>
+  <script type="module" src="/js/auth.js" defer></script>
 </head> 
 <body>
   <main class="container">
